The statistics buffer needs to be initialized with NaN so that inspecting statistics can have meaningful results

+++ components/statistic/src/main/java/org/limewire/statistic/NumericalStatistic.java
@@ -22,7 +22,7 @@ public abstract class NumericalStatistic extends AbstractStatistic {
         super.addData(data);
         synchronized(_buffer) {
             initializeBuffer();
-            _buffer.addLast(data);
+            _buffer.addLast((double)data);
         }
         _totalStatsRecorded++;
         if(data > _max) {
